Updated MainFeed
Completed Search/Filter for MainFeed
