+++ cookbook/src/components/features/Navigation.js
@@ -1,8 +1,6 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import AppBar from '@material-ui/core/AppBar';
-import TextField from '@material-ui/core/TextField';
-import SearchIcon from '@material-ui/icons/Search';
 import Link from '@material-ui/core/Link';
 import Toolbar from '@material-ui/core/Toolbar';
 import NavLogoGrey from '../logos/NavLogoGrey';
@@ -13,15 +11,11 @@ import axios from 'axios'
 const useStyles = makeStyles(theme => ({
   toolbar: {
     display: 'flex',
-    justifyContent: 'space-around',
+    justifyContent: 'space-between',
     backgroundColor: '#F2F2F2',
     border: 'none',
   },
-  textField: {
-    paddingBottom: '1.5rem',
-    width: '25rem',
-    color: '#A4A4A4',
-  },
+
   links: {
     width: 300,
     display: 'flex',
@@ -46,24 +40,7 @@ const HomeLink = React.forwardRef((props, ref) => (
 
 export default function Navigation() {
     const classes = useStyles();
-    const [recipe, setRecipe] = React.useState([]);
-    const [query, setQuery] = React.useState('');
 
-    React.useEffect(() => {
-      axios
-      .get(`https://lambdacooks.herokuapp.com/api/recipes`)
-      .then(res => {
-        const data= res.data;
-        const result = data.filter(post =>
-          post.recipe.toLowerCase().includes(query.toLowerCase())
-        );
-        setRecipe(result);
-      });
-    }, [query]);
-    
-    const handleInputChange = e => {
-      setQuery(e.target.value);
-    };
 
   return (
     <div>
@@ -72,18 +49,7 @@ export default function Navigation() {
           <Link component={HomeLink}>
             <NavLogoGrey />
           </Link>
-          <form className={classes.container} noValidate autoComplete="off">
-            
-            <TextField
-              id="standard-search"
-              label={<SearchIcon></SearchIcon>}
-              type="search"
-              onChange={handleInputChange}
-              value={query}
-              className={classes.textField}
-              margin="normal"
-            />
-          </form>
+          
           <div className={classes.links}>
               <RouterLink to="/sign-in" color="inherit" className={classes.btn}>Sign In</RouterLink>
               <RouterLink to="/sign-up" color="inherit" className={classes.btn}>Sign Up</RouterLink>
